feat(lumi-block): 添加AI润色和多语言切换功能
