+++ packages/apps/luminote.napp/lumi/lumi-page.html
@@ -12,6 +12,7 @@
       position: relative;
       display: block;
       margin: 0 auto;
+      padding: 0 32px;
       user-select: auto;
       /* background-color: rgb(62, 177, 142); */
     }
